Better validity-checking of namespaced symbols.
Before, namespaced symbols could have illegal characters which would
silently end symbol parsing and pass the extra characters back on the
Lisp reader.  Where these characters are not obvious indicators of the
end of a symbol (white space, parenthesis etc), OCML now signals an
error to the user who may have expected those characters to have been
parsed as part of the symbol.

Example: #_foo? would have parsed as two tokens, '#_foo' and '?',
where the user expected it to be a single token.