Fix the redefinition of ontologies.
This was broken in commit 768dcd24486c3707371ac95bb44f57ce9da76054.  A
well intentioned use of with-ontology meant that mutation of the OCML
special variables (the ones that get bound with
with-ocml-thread-safety) happened to the wrong variables: those still
associated with the old copy of ontology-directory.  When an ontology
is redefined, the ocml-ontology object retains its identity, but its
ontology-directory is swapped for an entirely new one.  We call
switch-to-ontology explicitly to bring the new ontology-directory into
play.