Pretty print namespaced symbols (Lispworks only)
That is, when printing a symbol that was read in as ‘#_foo:bar’,
instead of printing the expanded form ‘|http://www.foo.com/#bar|’, we
print ‘#_goo:bar’ where ‘goo’ is the currently active prefix for the
namespace ‘http://www.foo.com/#’.

This is implemented by defining an :around method for print-object,
something that not all Lisp systems support.  It is implemented here
only for Lispworks.