Generalise ontology content loading.
When def-ontology was read, it eventually led to the files listed in
the form's :file argument being loaded during ‘finalize-ontology’.
When we define an ontology from a live API instead of a file, we don't
have files to load, but may already have access to the ontology's
content.  Instead of hard-coding the loading of files, we now pass to
finalize-ontology a thunk which will call the define-class and
define-instance functions or load files during the define-ontology
call.  The file loading itself is now implemented as a special case of
this, through the function ‘load-files-thunk’.