The environment argument to #'unify is now non-optional
The environment argument should always be explicit.  When not
supplied, the default value was nil.  Unifying in an empty environment
is unusual thing to do: anyone wanting to do it should say so
explicitly.  It is far too easy to forget to supply a non-empty
environment.  Presumably there will be a small but tangible
performance benefit, since optional arguments are also inefficient.